refactor: client 패키지 생성, API 호출 helper 메서드 이동, 컴파일 오류 수정
